requirements.txt
openai
Twilio
Flask
python-dotenv==0.21.0